Backend/requirements.txt
fastapi==0.100.0  # FastAPI framework
pydantic==1.10.9  # For data validation and settings management
pandas==2.1.1  # For data manipulation and analysis
python-dotenv==1.0.0  # To load environment variables from a .env file
motor==3.2.0  # Asynchronous MongoDB driver
scikit-learn==1.3.1  # For machine learning models (pickle dependency)
uvicorn==0.23.0  # ASGI server to run FastAPI applications
